refactor(fischer): moves test directory one level up
required adding its own tsconfig to not mess with fischer's src's transpilation
