+++ fischer/__tests__/tsconfig.json
@@ -1,0 +1,43 @@
+{
+  "include": ["./**/*"],
+  "exclude": ["./*.js"],
+  "references": [{ "path": ".." }],
+  "compilerOptions": {
+    "incremental": true,
+    "module": "commonjs",
+    "lib": [
+      "ES5",
+      "ES6",
+      "DOM",
+      "ES2017",
+      "ES2018",
+      "ES2019",
+      "ES2020"
+    ],
+    "sourceMap": true,
+    "rootDir": ".",
+    "composite": true,
+    "removeComments": true,
+    "importHelpers": true,
+    "downlevelIteration": true,
+    "strict": true,
+    "noImplicitAny": true,
+    "strictNullChecks": true,
+    "strictFunctionTypes": true,
+    "alwaysStrict": true,
+    "noImplicitReturns": true,
+    "noUncheckedIndexedAccess": true,
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "allowJs": true,
+    "allowSyntheticDefaultImports": true,
+    "noFallthroughCasesInSwitch": true,
+    "resolveJsonModule": true,
+    "noEmit": true
+  },
+  "ts-node": {
+    "transpileOnly": true
+  }
+}
