+++ fischer/__tests__/ws/__utils__/socket.test.utils.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line node/no-unpublished-import
 import { io, Socket } from 'socket.io-client';
-import config from '../../../config/config';
+import config from '../../../src/config/config';
 
 const websocketPort = config.get('WEBSOCKET_PORT');
 let sockets: { [id: string]: Socket } = {};
