+++ fischer/__tests__/ws/connect.test.ts
@@ -1,4 +1,4 @@
-import { io } from '../../app';
+import { io } from '../../src/app';
 import * as socketUtils from './__utils__/socket.test.utils';
 
 const testUsername = 'brido';
